13/06/2023

+++ Assets/Scripts/GameController/ScoreCenter.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public static class ScoreCenter
 {
-    public static int score;
+    public static float score;
 
     public static void Reset()
     {
